requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch>=1.12.0
torchvision>=0.13.0
tensorboard==2.9.1
torch-tb-profiler==0.4.0
PyYAML==6.0
numpy==1.21.6
pandas
hdf5plugin==3.2.0
h5py==3.6.0
opencv-python==4.5.5.64
matplotlib==3.5.1
progress==1.6
mlflow==1.24.0
